change args order

+++ project2/stitching/src/cylindrical_wrap.rs
@@ -10,12 +10,12 @@ pub fn transform(img: &DynamicImage, f: f64) -> DynamicImage {
             let y_shift = (img.height() / 2) as f64;
             let x = col as f64 - x_shift;
             let y = row as f64 - y_shift;
-            let x_p = ((f * (x / f).tan()) + x_shift) as u32;
-            let y_p = (f * (y / f) / (x / f).cos() + y_shift) as u32;
-            if x_p < img.width() && y_p < img.height()
+            let x_p = ((f * (x / f).tan()) + x_shift) as i32;
+            let y_p = (f * (y / f) / (x / f).cos() + y_shift) as i32;
+            if (x_p as u32) < img.width() && (y_p as u32) < img.height()
                 && x_p >= 0 && y_p >= 0
             {
-                target.put_pixel(col, row, img.get_pixel(x_p, y_p));
+                target.put_pixel(col, row, img.get_pixel(x_p as u32, y_p as u32));
             }
             else {
                 target.put_pixel(col, row, Rgba { data: [0, 0, 0, 255] });
